pyproject.toml
[tool.poetry]
name = "rosenv"
version = "0.5.1a0"
description = "Management tool for ROS1 dependencies and package building."
authors = ["SSE Team <sse@honda-ri.eu>"]
readme = "README.md"

packages = [{ include = "rosenv", from = "src" }]

[tool.poetry.scripts]
rosenv = "rosenv.cli:main"

[[tool.poetry.source]]
name = "hri-gitlab"
url = "https://dmz-gitlab.honda-ri.de/api/v4/projects/502/packages/pypi/simple"
priority = "explicit"

[tool.poetry.dependencies]
python = "^3.8"
bloom = "^0.11.2"
cleo = "^2.0.1"
shellingham = "^1.5.0.post1"
pexpect = "^4.8.0"
defusedxml = "^0.7.1"
pyyaml = "^6.0"
deb-pkg-tools = "^8.4"

# we need this old version because of the option --install-layout for the makefile generated by bloom
setuptools = "45.2.0"

importlib-resources = { version = "^5.12.0", python = "<3.9" }
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.289"
black = "^23.3.0"
mypy = "^1.3.0"
pre-commit = "^3.3.1"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
types-pyyaml = "^6.0.12.10"
types-requests = "^2.31.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line_length = 120

[tool.mypy]
python_version = "3.8"
files = "src"

strict = true

warn_return_any = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true

disallow_untyped_defs = true
no_implicit_optional = true

show_error_context = true
show_error_codes = true
show_error_code_links = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "pexpect",
    "pexpect.*",
    "shellingham",
    "defusedxml",
    "deb_pkg_tools.*",
    "importlib_resources.*",
]
ignore_missing_imports = true

[tool.ruff]
select = ["ALL", "PLR6301"]
ignore = [
    "D1",      # docstrings
    "D203",    # one-blank-line-before-class
    "D212",    # multi-line-summary-first-line
    "ERA001",  # Commented out code found
    "PLC0414", # useless import alias
    "TCH",     # type checking blocks
    "ANN10",   # typing self and cls
    "FIX002",  # TODO anywhere in comments
]

line-length = 120
target-version = "py38"

[tool.ruff.per-file-ignores]
"src/*" = ["PT"] # pytest rule set
"src/rosenv/commands/*" = ["RUF012"]
"tests/*" = [
    "D1",      # docstrings
    "T20",     # print detection
    "S101",    # `assert` detected
    "PLR0913", # too-many-arguments
]
"pkgInfo.py" = [
    "D100", # public docstring
    "N816", # mixed case variables
]
"tests/resources/*" = ["ALL"]

[tool.ruff.isort]
force-single-line = true
lines-between-types = 1
lines-after-imports = 2
known-first-party = ["rosenv"]
required-imports = ["from __future__ import annotations"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
enable     = [ 'E' ]
disable    = [ 'W', 'C', 'R', 'E0602', 'import-error', 'no-member', 'no-name-in-module']

